Small cleanup - separate views into api and views, bg gradient
