--- buildings/templates/buildings/all_buildings.html
@@ -1,100 +1,0 @@
-
-{% extends "buildings/base.html" %}
-
-
-{% block content %}
-
-<div class="container">
-    <form action="{% url 'buildings:all_buildings' %}" class="buildings-search-query" method="GET" autocomplete="off">
-        <p class="mb-3">Search filters:</p>
-        <div class="building-search-filters">
-            <div class="input-group mb-3 me-3">
-                <div class="input-group-prepend">
-                    <span class="input-group-text">Address</span>
-                </div>
-                <input type="text" name="q_address" class="q_address" value="{{ request.GET.q_address }}" 
-                    size="18"
-                    hx-get="{% url 'buildings:all_buildings' %}" 
-                    hx-trigger="keyup changed delay:200ms"
-                    hx-include=".buildings-search-query"
-                    hx-target="#building-search-target"
-                    />
-            </div>
-            
-            <div class="input-group mb-3 me-3">
-                <div class="input-group-prepend">
-                    <span class="input-group-text">Locality</span>
-                </div>
-                <input type="text" name="q_locality" class="q_locality" value="{{ request.GET.q_locality }}" 
-                    size="18"
-                    hx-get="{% url 'buildings:all_buildings' %}" 
-                    hx-trigger="keyup changed delay:200ms"
-                    hx-include=".buildings-search-query"
-                    hx-target="#building-search-target"
-                    />
-            </div>
-            <div class="input-group mb-3 me-3" >
-                <div class="input-group-prepend">
-                    <span class="input-group-text">Region</span>
-                </div>
-                <input type="text" name="q_region" class="q_region" value="{{ request.GET.q_region }}" 
-                    size="18"
-                    hx-get="{% url 'buildings:all_buildings' %}" 
-                    hx-trigger="keyup changed delay:200ms"
-                    hx-include=".buildings-search-query"
-                    hx-target="#building-search-target"
-                    />
-            </div>
-            <div class="input-group mb-3 me-3">
-                <div class="input-group-prepend">
-                    <span class="input-group-text">CUBF</span>
-                </div>
-                <input type="text" name="q_cubf" class="q_cubf" value="{{ request.GET.q_cubf }}" 
-                    size="6"
-                    hx-get="{% url 'buildings:all_buildings' %}" 
-                    hx-trigger="keyup changed delay:200ms"
-                    hx-include=".buildings-search-query"
-                    hx-target="#building-search-target"
-                    />
-            </div>
-            
-            <div class="input-group mb-3 me-3">
-                <div class="input-group-prepend">
-                    <span class="input-group-text">Votes</span>
-                </div>
-                <select name="q_num_votes_op" hx-include=".buildings-search-query">
-                    <option value="gte" {% if request.GET.q_num_votes_op == 'gte' or request.GET.q_num_votes_op is None %}selected{% endif %}>&ge;</option>
-                    <option value="lte" {% if request.GET.q_num_votes_op == 'lte' %}selected{% endif %}>&le;</option>
-                    <option value="gt" {% if request.GET.q_num_votes_op == 'gt' %}selected{% endif %}>&gt;</option>
-                    <option value="lt" {% if request.GET.q_num_votes_op == 'lt' %}selected{% endif %}>&lt;</option>
-                    <option value="eq" {% if request.GET.q_num_votes_op == 'eq' %}selected{% endif %}>=</option>
-                </select>
-                <input type="number" name="q_num_votes" value="{{ request.GET.q_num_votes }}"
-                    min="0"
-                    hx-get="{% url 'buildings:all_buildings' %}" 
-                    hx-trigger="keyup changed delay:200ms"
-                    hx-include=".buildings-search-query"
-                    hx-target="#building-search-target"
-                    />
-            </div>
-            <div class="input-group mb-3 me-3">
-                <input type="submit" value="Search"/>
-            </div>
-        </div>
-    </form>
-    <div class="mb-1"> 
-        {{ page_obj.paginator.count }} results
-    </div>
-    <div>
-        {% if request.GET.q %}
-            Search results for {{ request.GET.q }} under {{ request.GET.type }}
-        {% endif %}
-    </div>
-
-
-    {% include 'buildings/partials/all_buildings.html' %}
-
-</div>
-
-
-{% endblock content %}
