+++ buildings/templates/buildings/base.html
@@ -77,8 +77,12 @@
         <div class="absolute inset-0 -z-[1] overflow-hidden bg-blue-100">
           <div class="jumbo transform-gpu absolute -inset-[10px] opacity-30 -z-[1]">
           </div>
+          <div class="absolute inset-x-0 -z-[10] top-[calc(75%)] transform-gpu overflow-hidden blur-[100px]" aria-hidden="true">
+              <div class="aspect-[3/2] w-[50%] -translate-x-1/2 bg-gradient-to-tr from-[#a279d1] to-[#a279d1] opacity-30  -z-[1]"
+                style="clip-path: polygon(4.1% 44.1%, 100% 6.6%, 97.5% 2.9%, 85.5% 0.1%, 80.7% 2%, 72.5% 32.5%, 60.2% 62.4%, 52.4% 68.1%)">
+              </div>
+          </div>
         </div>
-
 
         <header id="header" class="z-[10]">
             {% include 'buildings/navbar.html' %}
