+++ buildings/templates/buildings/surveys/survey_partial.html
