+++ buildings/views/__init__.py
@@ -1,0 +1,2 @@
+from .views import *
+from .api import *
